split montage-(deserializer|serializer) element into sperate specs
